Step2: Generating Data (Resembling Rats)
